requirements.txt
groq
pydantic
streamlit
python-dotenv
PyPDF2